# specs/001-fitness-tracker-app/contracts/api.yaml
openapi: 3.1.0
info:
  title: Fitness Tracker API
  description: |
    Backend API for the Fitness Tracker application.
    Provides endpoints for workout management, session logging, and Google Fit integration.
  version: 1.0.0
  contact:
    name: Fitness Tracker Team
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://fitness-tracker-api.web.app
    description: Production (Firebase)
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/users/me:
    get:
      tags: [users]
      summary: Get current user profile
      operationId: getCurrentUser
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: User not found (first login, needs initialization)

    put:
      tags: [users]
      summary: Update user profile
      operationId: updateCurrentUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/users/me/initialize:
    post:
      tags: [users]
      summary: Initialize new user with default workout plan
      description: Called on first login to seed user data
      operationId: initializeUser
      responses:
        '201':
          description: User initialized with default plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: User already initialized

  # ============================================================================
  # WORKOUT PLAN ENDPOINTS
  # ============================================================================
  /api/v1/workouts/plan:
    get:
      tags: [workouts]
      summary: Get active workout plan
      operationId: getActivePlan
      responses:
        '200':
          description: Active workout plan with all days
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutPlanFull'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No active plan found

    put:
      tags: [workouts]
      summary: Update workout plan metadata
      operationId: updatePlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutPlanUpdate'
      responses:
        '200':
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/workouts/plan/reset:
    post:
      tags: [workouts]
      summary: Reset plan to default
      description: Replaces current plan with the default knee-preservation program
      operationId: resetPlanToDefault
      responses:
        '200':
          description: Plan reset to default
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutPlanFull'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/workouts/today:
    get:
      tags: [workouts]
      summary: Get today's workout
      description: Returns the workout for the current day of the week
      operationId: getTodayWorkout
      responses:
        '200':
          description: Today's workout with exercises
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutDayFull'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No workout defined for today
  /api/v1/workouts/days/{dayId}:
    get:
      tags: [workouts]
      summary: Get specific workout day
      operationId: getWorkoutDay
      parameters:
        - name: dayId
          in: path
          required: true
          schema:
            type: string
            enum: [sunday, monday, tuesday, wednesday, thursday, friday, saturday]
      responses:
        '200':
          description: Workout day with exercises
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutDayFull'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Workout day not found

    put:
      tags: [workouts]
      summary: Update workout day
      operationId: updateWorkoutDay
      parameters:
        - name: dayId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutDayUpdate'
      responses:
        '200':
          description: Updated workout day
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutDay'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/workouts/days/{dayId}/exercises:
    get:
      tags: [workouts]
      summary: List exercises for a workout day
      operationId: listExercises
      parameters:
        - name: dayId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Exercise'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [workouts]
      summary: Add exercise to workout day
      operationId: addExercise
      parameters:
        - name: dayId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseCreate'
      responses:
        '201':
          description: Created exercise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/workouts/days/{dayId}/exercises/{exerciseId}:
    put:
      tags: [workouts]
      summary: Update exercise
      operationId: updateExercise
      parameters:
        - name: dayId
          in: path
          required: true
          schema:
            type: string
        - name: exerciseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseUpdate'
      responses:
        '200':
          description: Updated exercise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Exercise not found

    delete:
      tags: [workouts]
      summary: Remove exercise from workout day
      operationId: deleteExercise
      parameters:
        - name: dayId
          in: path
          required: true
          schema:
            type: string
        - name: exerciseId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Exercise deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Exercise not found

  # ============================================================================
  # SESSION ENDPOINTS
  # ============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token
  responses:
    Unauthorized:
      description: Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    User:
      type: object
      required: [uid, email, createdAt, preferences, stats]
      properties:
        uid:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
        createdAt:
          type: string
          format: date-time
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        googleFit:
          $ref: '#/components/schemas/GoogleFitStatus'
        stats:
          $ref: '#/components/schemas/UserStats'
    UserUpdate:
      type: object
      properties:
        displayName:
          type: string
        preferences:
          $ref: '#/components/schemas/UserPreferences'
    UserPreferences:
      type: object
      properties:
        showWarmup:
          type: boolean
          default: true
        darkMode:
          type: boolean
          default: true
        weekStartsOn:
          type: integer
          enum: [0, 1]
          default: 1
    UserStats:
      type: object
      properties:
        totalWorkouts:
          type: integer
        currentStreak:
          type: integer
        longestStreak:
          type: integer
        lastWorkoutAt:
          type: string
          format: date-time
    GoogleFitStatus:
      type: object
      properties:
        connected:
          type: boolean
        lastSyncAt:
          type: string
          format: date-time

    # --------------------------------------------------------------------------
    # WORKOUT PLAN SCHEMAS
    # --------------------------------------------------------------------------
    WorkoutPlan:
      type: object
      required: [id, name, isActive]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkoutPlanFull:
      allOf:
        - $ref: '#/components/schemas/WorkoutPlan'
        - type: object
          properties:
            days:
              type: array
              items:
                $ref: '#/components/schemas/WorkoutDayFull'
    WorkoutPlanUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string

    # --------------------------------------------------------------------------
    # WORKOUT DAY SCHEMAS
    # --------------------------------------------------------------------------
    WorkoutDay:
      type: object
      required: [id, dayOfWeek, name, isRestDay]
      properties:
        id:
          type: string
        dayOfWeek:
          type: integer
          minimum: 0
          maximum: 6
        name:
          type: string
        focus:
          type: string
        format:
          type: string
        duration:
          type: integer
          description: Estimated duration in minutes
        isRestDay:
          type: boolean
        sortOrder:
          type: integer
    WorkoutDayFull:
      allOf:
        - $ref: '#/components/schemas/WorkoutDay'
        - type: object
          properties:
            exercises:
              type: array
              items:
                $ref: '#/components/schemas/Exercise'
    WorkoutDayUpdate:
      type: object
      properties:
        name:
          type: string
        focus:
          type: string
        format:
          type: string
        duration:
          type: integer
        isRestDay:
          type: boolean

    # --------------------------------------------------------------------------
    # EXERCISE SCHEMAS
    # --------------------------------------------------------------------------
    Exercise:
      type: object
      required: [id, name, instructions, sortOrder]
      properties:
        id:
          type: string
        name:
          type: string
        instructions:
          type: string
        sets:
          type: integer
        reps:
          oneOf:
            - type: integer
            - type: string
        tempo:
          type: string
        notes:
          type: string
        sortOrder:
          type: integer
        category:
          type: string
    ExerciseCreate:
      type: object
      required: [name, instructions]
      properties:
        name:
          type: string
        instructions:
          type: string
        sets:
          type: integer
        reps:
          oneOf:
            - type: integer
            - type: string
        tempo:
          type: string
        notes:
          type: string
        category:
          type: string
    ExerciseUpdate:
      type: object
      properties:
        name:
          type: string
        instructions:
          type: string
        sets:
          type: integer
        reps:
          oneOf:
            - type: integer
            - type: string
        tempo:
          type: string
        notes:
          type: string
        sortOrder:
          type: integer
        category:
          type: string

    # --------------------------------------------------------------------------
    # SESSION SCHEMAS
    # --------------------------------------------------------------------------
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object